requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Flask-Mail==0.9.1
Werkzeug==2.3.7
SQLAlchemy==2.0.20
gunicorn==21.2.0
psycopg2-binary==2.9.7
